+++ Demo/CqlSdkUseCases/ProgramInputArgs.cs
@@ -1,0 +1,3 @@
+﻿namespace CqlSdkUseCases;
+
+public record ProgramInputArgs(string Name, string ElmDir, string CSharpDir, string DllDir, Func<FileInfo, bool>? FileFilter = null);
